Register Steps[1-4]-WIP
