+++ src/Routing.jsx
@@ -7,7 +7,11 @@ import { Routes, Route, BrowserRouter as Router } from "react-router-dom";
 const Home = lazy(() => import("./pages/LandingPage"));
 const MapList = lazy(() => import("./pages/ListMapPage"));
 const Canvas = lazy(() => import("./pages/MapCanvasPage"));
-const Auth = lazy(()=>import("./pages/Auth"));
+const Auth = lazy(()=>import("./pages/authentication/Auth"));
+const ContactDetails = lazy(()=>import("./pages/authentication/ContactDetails"));
+const EmailVerification = lazy(()=>import("./pages/authentication/EmailVerification"));
+const AccountCredentials = lazy(()=>import("./pages/authentication/AccountCredentials"));
+
 
 function Routing() {
   // TODO once the auth is ready check if the user is connected
@@ -22,6 +26,9 @@ function Routing() {
           <Route path="/dashboard/:mapId" element={<Canvas />} exact />
           <Route path="/dashboard/:mapId" element={<Canvas />} exact />
           <Route path="/authentication" element={<Auth />} exact />
+          <Route path="/contactdetails" element={<ContactDetails />} exact />
+          <Route path="/emailverification" element={<EmailVerification/>} exact />
+          <Route path="/credentials" element={<AccountCredentials/>} exact />
           <Route path="*" element={<div>404 Not found</div>} />
         </Routes>
       </Router>
