+++ src/pages/authentication/EmailConfrimation.jsx
@@ -1,0 +1,11 @@
+import React, { Component } from 'react'
+
+export class EmailConfrimation extends Component {
+  render() {
+    return (
+      <div>EmailConfrimation</div>
+    )
+  }
+}
+
+export default EmailConfrimation
