+++ src/pages/authentication/SignIn.jsx
@@ -1,0 +1,221 @@
+import React, { Component } from 'react';
+import { Container } from '@chakra-ui/react';
+import styled from "styled-components";
+import {  Box, Flex,  Checkbox, Button, InputGroup, Input, Stack, InputLeftElement } from '@chakra-ui/react';
+import {Tag, Unlock} from '@styled-icons/bootstrap';
+
+const ForgotMyUsername = styled.section`
+    width: 274px;
+height: 20px;
+font-family: 'Lucida Sans';
+font-style: normal;
+font-weight: 400;
+font-size: 14px;
+line-height: 20px;
+display: flex;
+align-items: center;
+text-align: center;
+letter-spacing: 0.02em;
+
+
+color: #2A69AC;
+flex: none;
+order: 0;
+flex-grow: 1;
+`
+
+const FooterText = styled.section`
+width: 564px;
+height: 20px;
+justify-content: center;
+font-family: 'Lucida Sans';
+font-style: normal;
+font-weight: 400;
+font-size: 14px;
+line-height: 20px;
+display: flex;
+align-items: center;
+text-align: center;
+letter-spacing: 0.02em;
+color: #001011;
+flex: none;
+order: 0;
+flex-grow: 0;
+`
+
+const RememberMe = styled.section`
+    width: 204px;
+height: 20px;
+font-family: 'Lucida Sans';
+font-style: normal;
+font-weight: 400;
+font-size: 14px;
+line-height: 20px;
+letter-spacing: 0.02em;
+color: #001011;
+flex: none;
+order: 1;
+flex-grow: 0;
+top:10px;
+`
+
+const SignInWrapper = styled.section `
+    width: 59px;
+height: 24px;
+
+font-family: 'Lucida Sans';
+font-style: normal;
+font-weight: 600;
+font-size: 16px;
+line-height: 24px;
+letter-spacing: 0.02em;
+color: #FFFFFF;
+flex: none;
+order: 0;
+flex-grow: 0;
+`
+
+const FormFieldWrapper = styled.section`
+width: 564px;
+height: 20px;
+font-family: 'Lucida Sans';
+font-style: normal;
+font-weight: 400;
+font-size: 14px;
+line-height: 20px;
+letter-spacing: 0.02em;
+margin-top:24px;
+/* margin-left:40px; */
+color: #001011;
+flex: none;
+order: 0;
+align-self: stretch;
+flex-grow: 0;
+`
+const Username = styled.section`
+width: 564px;
+height: 24px;
+
+margin-top:24px;
+/* margin-left:40px; */
+font-family: 'Lucida Sans';
+font-style: normal;
+font-weight: 400;
+font-size: 16px;
+line-height: 24px;
+letter-spacing: 0.02em;
+color: #001011;
+
+flex: none;
+order: 0;
+align-self: stretch;
+flex-grow: 0;
+`
+
+const Explanation = styled.section`
+width: 564px;
+height: 16px;
+font-family: 'Lucida Sans Unicode';
+font-style: normal;
+font-weight: 400;
+font-size: 12px;
+line-height: 16px;
+letter-spacing: 0.01em;
+color: #718096;
+margin-top: 8px;
+flex: none;
+order: 2;
+align-self: stretch;
+flex-grow: 0;
+`
+
+export class SignIn extends Component {
+  render() {
+    return (
+        <Container maxW='container.lg' marginTop='35px'>
+             <FormFieldWrapper>
+        Sign in with your CirclePass account to access our apps and services
+    </FormFieldWrapper>
+    <Box>
+        <Box>
+        <Username color='#001011' size='16px' lineHeight='24px'>
+            Username
+        </Username>
+            <InputGroup>
+                <InputLeftElement
+                pointerEvents='none'
+                ><Tag
+                size="20"
+                color='#A3A3A3'
+                />
+        </InputLeftElement>
+            <Input  width='90%'  bgColor='#ffffff' type='Username' placeholder='    Username' />
+        </InputGroup>
+        <Explanation>
+            Enter the username you have registered with
+        </Explanation>
+        </Box>
+        <Box>
+        <Username color='#001011' size='16px' lineHeight='24px'>
+            Password
+        </Username>
+        <InputGroup>
+        <InputLeftElement
+            pointerEvents='none'
+            ><Unlock
+            color='#A3A3A3'
+            size="20"
+            />
+        </InputLeftElement>
+            <Input width='90%'  bgColor='#ffffff' type='Password' placeholder='    Password' />
+        </InputGroup>
+        <Explanation>
+            Enter the password you have registered with
+        </Explanation>
+       </Box>
+       <Stack marginTop='15px'>
+        <Checkbox colorScheme='blue' defaultChecked>
+            <RememberMe>
+                Remember my on this device
+            </RememberMe>
+        </Checkbox>
+        </Stack>
+        <Stack display='flex' justifyContent='center' spacing={4} direction='row' align='center'>
+            <Button 
+            top='30px'
+            display='flex' 
+            flex-direction= 'row'
+            justify-content= 'center'
+            align-items='center'
+            padding='10px 16px'
+            gap='8px'
+            width= '91px'
+            _hover={{ bg: "#00A0E9" }}
+            height= '40px'
+            background= '#00A0E9' 
+            border-radius='6px'
+            >
+            <SignInWrapper>
+                Sign in
+            </SignInWrapper>
+            </Button>
+        </Stack>
+        {/* Some changes to be done... */}
+        <Box bottom={'30px'} position='absolute'
+        marginTop={'230px'} marginLeft={'150px'}
+        >
+            <FooterText>
+            Having a problem signing in?
+            </FooterText>
+            <Flex  marginLeft={'85px'} marginTop={'10px'} flexDirection={'row'}>
+                <ForgotMyUsername>Forgot my username &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp; | &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;Reset my password</ForgotMyUsername>                                            
+            </Flex>
+        </Box>
+    </Box>
+        </Container>
+       
+    )
+  }
+}
+
+export default SignIn
