+++ src/pages/authentication/VerificationCompleted.jsx
@@ -1,0 +1,11 @@
+import React, { Component } from 'react'
+
+export class VerificationCompleted extends Component {
+  render() {
+    return (
+      <p>VerificationCompleted</p>
+    )
+  }
+}
+
+export default VerificationCompleted
